requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
jinja2==2.8
markupsafe==0.23          # via jinja2
pillow==3.0.0
selenium==2.48.0
